+++ packages/web/src/containers/profile-modal/createChatButton.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/navigation'
 
 import { Button } from '@packages/ui'
 
-import { createOnetoOneChat } from '@/apis/createOnetoOneChat'
+import { createOnetoOneChat } from '@/apis/createChat'
 
 export default function CreateChatButton({ uuid }: { uuid: string }) {
   const router = useRouter()
